apps/llm-sampler/pyproject.toml
[project]
name = "llm-sampler"
version = "0.1.0"
description = "Minimal Streamlit + FastAPI app to try LLMs (managed with uv)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "streamlit>=1.37.0",
    "python-dotenv>=1.0.1",
    "openai>=1.40.0",
]

[tool.uv]
# No special config needed; uv will read this file to resolve deps.
